requirements.txt
django==3.1.3
djangorestframework==3.12.2
black==20.8b1
isort==5.6.4
